xfree86: vgahw: move over public SDK header to include/
